requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
openai==1.63.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
